backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
supabase
google-generativeai
pydantic
pydantic-settings
httpx
PyMuPDF
pathlib
